organized some of the TX concurrency tests. still WIP.
