+++ test/src/test/java/org/corfudb/runtime/object/transactions/AbstractTransactionContextTest.java
@@ -3,17 +3,24 @@ package org.corfudb.runtime.object.transactions;
 import com.google.common.reflect.TypeToken;
 import org.corfudb.runtime.collections.ISMRMap;
 import org.corfudb.runtime.collections.SMRMap;
+import org.corfudb.runtime.exceptions.TransactionAbortedException;
+import org.corfudb.runtime.object.CorfuSharedCounter;
 import org.corfudb.runtime.view.AbstractViewTest;
 import org.junit.Before;
+import org.junit.Test;
 
+import java.util.ArrayList;
 import java.util.Map;
+import java.util.concurrent.atomic.AtomicIntegerArray;
+
+import static org.assertj.core.api.Assertions.fail;
 
 /**
  * Created by mwei on 11/21/16.
  */
 public abstract class AbstractTransactionContextTest extends AbstractViewTest {
 
-    private ISMRMap<String, String> testMap;
+    protected ISMRMap<String, String> testMap;
 
     abstract void TXBegin();
 
@@ -51,5 +58,40 @@ public abstract class AbstractTransactionContextTest extends AbstractViewTest {
     // Write only, don't read previous value.
     void write(String key, String value) {
         getMap().blindPut(key, value);
+    }
+
+    // counter-array tests below
+
+    static final int INITIAL = 32;
+    static final int OVERWRITE_ONCE = 33;
+    static final int OVERWRITE_TWICE = 34;
+
+    int numTasks;
+    ArrayList<CorfuSharedCounter> sharedCounters;
+    AtomicIntegerArray commitStatus;
+    final int COMMITVALUE = 1; // by default, ABORTVALUE = 0
+    AtomicIntegerArray snapStatus = new AtomicIntegerArray(numTasks);
+
+    /**
+     * build an array of shared counters for the test
+     */
+    void setupCounters() {
+
+        numTasks = PARAMETERS.NUM_ITERATIONS_MODERATE;
+        sharedCounters = new ArrayList<>();
+
+        for (int i = 0; i < numTasks; i++)
+            sharedCounters.add(i, getRuntime().getObjectsView()
+                    .build()
+                    .setStreamName("test"+i)
+                    .setType(CorfuSharedCounter.class)
+                    .open() );
+
+        // initialize all shared counters
+        for (int i = 0; i < numTasks; i++)
+            sharedCounters.get(i).setValue(INITIAL);
+
+        commitStatus = new AtomicIntegerArray(numTasks);
+        snapStatus = new AtomicIntegerArray(numTasks);
     }
 }
